Inspecting Solutions

Inspecting Solutions

Tabulate Decision Rule

tabulate

Impulse Response Functions

Dolo.response — Function.

Function "response" computes the impulse response functions with several major options:

the user can provide a vector with the first values of the model's exogenous processes, e1.

the user can provide a name of the shock of interest and the size of the shock_name.

the user can provide only a name of the shock of interest. The size of the shock is assumed to be a one standard deviation given in the yaml file.

Arguments

model::NumericModel: Model object that describes the current model environment.

dr: Solved decision rule.

e1::ListOfPoints: List of initial model's exogenous processes values.

If e1 is not provided, then:

shock_name: the name of the shock of interest.

optional:

Impulse: the size of the shock; default: one standard deviation.

s0::ListOfPoints: List of initial state variable values; default: model.calibration[:states]

Returns

response: Impulse response function.

source

Stochastic Simulation

QuantEcon.simulate — Function.

This is the one we document.

source